requirements.txt
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-mistralai==0.1.6
langchain-text-splitters==0.0.1
chromadb==0.5.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1
tiktoken==0.7.0
lxml==5.2.2
sentence-transformers==2.7.0